+++ src/class/flagshipVisitor/flagshipVisitor.test.ts
@@ -1225,6 +1225,44 @@ describe('FlagshipVisitor', () => {
                 try {
                     expect(response).toEqual({
                         campaignId: 'bmjdprsjan0g01uq2ceg',
+                        isReference: false,
+                        variationGroupId: 'bmjdprsjan0g01uq2ceg',
+                        variationId: 'bmjdprsjan0g01uq1ctg'
+                    });
+                } catch (error) {
+                    done.fail(error);
+                }
+                done();
+            });
+            mockAxios.mockResponse(responseObj);
+            const url = `${internalConfig.apiV2}${demoData.envId[0]}/campaigns?mode=normal`;
+            expect(mockAxios.post).toHaveBeenNthCalledWith(
+                1,
+                url,
+                {
+                    context: demoData.visitor.cleanContext,
+                    trigger_hit: false,
+                    visitor_id: demoData.visitor.id[0]
+                },
+                {
+                    ...assertionHelper.getCampaignsQueryParams(),
+                    ...assertionHelper.getTimeout(url, sdk.config),
+
+                    ...assertionHelper.getApiKeyHeader(demoData.apiKey[0])
+                }
+            );
+        });
+        it('should return tell if the modification receive is a ref when it is the case', (done) => {
+            const responseObj = {
+                data: { ...demoData.decisionApi.normalResponse.manyModifInManyCampaigns },
+                status: 200,
+                statusText: 'OK'
+            };
+            visitorInstance.getModificationInfo('hello').then((response) => {
+                try {
+                    expect(response).toEqual({
+                        campaignId: 'bmjdprsjan0g01uq2ceg',
+                        isReference: true,
                         variationGroupId: 'bmjdprsjan0g01uq2ceg',
                         variationId: 'bmjdprsjan0g01uq1ctg'
                     });
@@ -1258,6 +1296,7 @@ describe('FlagshipVisitor', () => {
             visitorInstance.getModificationInfo('psp').then((response) => {
                 try {
                     expect(response).toEqual({
+                        isReference: false,
                         campaignId: 'blntcamqmdvg04g371f0',
                         variationGroupId: 'blntcamqmdvg04g371h0',
                         variationId: 'blntcamqmdvg04g371hg'
@@ -1997,10 +2036,10 @@ describe('FlagshipVisitor', () => {
                 statusText: 'OK'
             };
         });
-        it('should send hit of type "transaction" + "item" + "event" + "page" + "screen" if there are in the array argument', (done) => {
+        it('should send hit of type "transaction" + "item" + "event" + "pageview" + "screenview" if there are in the array argument', (done) => {
             try {
                 visitorInstance
-                    .sendHits([demoData.hit.event, demoData.hit.item, demoData.hit.screen, demoData.hit.transaction])
+                    .sendHits([demoData.hit.event, demoData.hit.item, demoData.hit.screenview, demoData.hit.transaction, demoData.hit.pageview])
                     .then((response) => {
                         try {
                             expect(response).not.toBeDefined();
@@ -2008,6 +2047,7 @@ describe('FlagshipVisitor', () => {
                             expect(spyInfoLogs).toHaveBeenNthCalledWith(2, 'sendHits - hit (type"ITEM") send successfully');
                             expect(spyInfoLogs).toHaveBeenNthCalledWith(3, 'sendHits - hit (type"SCREENVIEW") send successfully');
                             expect(spyInfoLogs).toHaveBeenNthCalledWith(4, 'sendHits - hit (type"TRANSACTION") send successfully');
+                            expect(spyInfoLogs).toHaveBeenNthCalledWith(5, 'sendHits - hit (type"PAGEVIEW") send successfully');
                             expect(spyWarnLogs).toBeCalledTimes(0);
                             expect(spyErrorLogs).toBeCalledTimes(0);
                             expect(spyFatalLogs).toBeCalledTimes(0);
@@ -2023,7 +2063,7 @@ describe('FlagshipVisitor', () => {
                 mockAxios.mockResponse();
                 mockAxios.mockResponse();
                 mockAxios.mockResponse();
-                // mockAxios.mockResponse()
+                mockAxios.mockResponse();
                 expect(mockAxios.post).toHaveBeenNthCalledWith(1, 'https://ariane.abtasty.com', {
                     cid: 'bn1ab7m56qolupi5sa0g',
                     vid: 'test-perf',
@@ -2076,7 +2116,15 @@ describe('FlagshipVisitor', () => {
                     ts: 888,
                     tt: 1234444
                 });
-                expect(mockAxios.post).toHaveBeenCalledTimes(4);
+                expect(mockAxios.post).toHaveBeenNthCalledWith(5, 'https://ariane.abtasty.com', {
+                    cid: 'bn1ab7m56qolupi5sa0g',
+                    dl: 'http%3A%2F%2Fabtastylab.com%2F60511af14f5e48764b83d36ddb8ece5a%2F',
+                    ds: 'APP',
+                    t: 'PAGEVIEW',
+                    vid: 'test-perf',
+                    pt: 'YoloPage'
+                });
+                expect(mockAxios.post).toHaveBeenCalledTimes(5);
             } catch (error) {
                 done.fail(error);
             }
@@ -2144,9 +2192,34 @@ describe('FlagshipVisitor', () => {
             });
             expect(mockAxios.post).toBeCalledTimes(0);
         });
-        it('should logs error when hit "screen" not set correctly', (done) => {
-            const brokenHit1 = { ...demoData.hit.screen, data: { ...demoData.hit.screen.data, pageTitle: null } };
-            const brokenHit2 = { ...demoData.hit.screen, data: { ...demoData.hit.screen.data, documentLocation: null } };
+        it('should logs error when hit "pageview" not set correctly', (done) => {
+            const brokenHit1 = { ...demoData.hit.pageview, data: { ...demoData.hit.pageview.data, pageTitle: null } };
+            const brokenHit2 = { ...demoData.hit.pageview, data: { ...demoData.hit.pageview.data, documentLocation: null } };
+            visitorInstance.sendHits([brokenHit1, brokenHit2]).then((response) => {
+                try {
+                    expect(response).not.toBeDefined();
+                    expect(spyInfoLogs).toBeCalledTimes(0);
+                    expect(spyWarnLogs).toBeCalledTimes(0);
+                    expect(spyErrorLogs).toHaveBeenNthCalledWith(
+                        1,
+                        'sendHits(PageView) - failed because following required attribute "pageTitle" is missing...'
+                    );
+                    expect(spyErrorLogs).toHaveBeenNthCalledWith(
+                        2,
+                        'sendHits(PageView) - failed because following required attribute "documentLocation" is missing...'
+                    );
+                    expect(spyErrorLogs).toBeCalledTimes(2);
+                    expect(spyFatalLogs).toBeCalledTimes(0);
+                } catch (error) {
+                    done.fail(error);
+                }
+                done();
+            });
+            expect(mockAxios.post).toBeCalledTimes(0);
+        });
+        it('should logs error when hit "screenview" not set correctly', (done) => {
+            const brokenHit1 = { ...demoData.hit.screenview, data: { ...demoData.hit.screenview.data, pageTitle: null } };
+            const brokenHit2 = { ...demoData.hit.screenview, data: { ...demoData.hit.screenview.data, documentLocation: null } };
             visitorInstance.sendHits([brokenHit1, brokenHit2]).then((response) => {
                 try {
                     expect(response).not.toBeDefined();