+++ src/class/flagshipVisitor/flagshipVisitor.ts
@@ -32,6 +32,9 @@ import {
     ActivatedArchived,
     UnauthenticateVisitorOutput,
     AuthenticateVisitorOutput
+    ActivatedArchived,
+    ScreenViewHit,
+    PageViewHit
 } from './types';
 
 class FlagshipVisitor extends EventEmitter implements IFlagshipVisitor {
@@ -579,6 +582,7 @@ class FlagshipVisitor extends EventEmitter implements IFlagshipVisitor {
         }
         const polishOutput = (data: DecisionApiCampaign): GetModificationInfoOutput => ({
             campaignId: data.id,
+            isReference: !!data.variation.reference,
             variationId: data.variation.id,
             variationGroupId: data.variationGroupId
         });
@@ -924,8 +928,7 @@ class FlagshipVisitor extends EventEmitter implements IFlagshipVisitor {
         const optionalAttributes: { [key: string]: string | number | boolean } = {};
         // TODO: move common optional attributes before switch statement (ie: "pageTitle", "documentLocation",...)
         switch (hitData.type.toUpperCase()) {
-            case 'SCREEN':
-            case 'SCREENVIEW': {
+            case 'SCREEN': {
                 const { documentLocation, pageTitle } = hitData.data;
                 if (!documentLocation || !pageTitle) {
                     if (!documentLocation)
@@ -938,6 +941,40 @@ class FlagshipVisitor extends EventEmitter implements IFlagshipVisitor {
                 }
                 return {
                     t: 'SCREENVIEW',
+                    dl: documentLocation,
+                    pt: pageTitle
+                };
+            }
+            case 'SCREENVIEW': {
+                const { documentLocation, pageTitle } = hitData.data as ScreenViewHit;
+                if (!documentLocation || !pageTitle) {
+                    if (!documentLocation)
+                        this.log.error(
+                            'sendHits(ScreenView) - failed because following required attribute "documentLocation" is missing...'
+                        );
+                    if (!pageTitle)
+                        this.log.error('sendHits(ScreenView) - failed because following required attribute "pageTitle" is missing...');
+                    return null;
+                }
+                return {
+                    t: 'SCREENVIEW',
+                    dl: documentLocation,
+                    pt: pageTitle
+                };
+            }
+            case 'PAGEVIEW': {
+                const { documentLocation, pageTitle } = hitData.data as PageViewHit;
+                if (!documentLocation || !pageTitle) {
+                    if (!documentLocation)
+                        this.log.error(
+                            'sendHits(PageView) - failed because following required attribute "documentLocation" is missing...'
+                        );
+                    if (!pageTitle)
+                        this.log.error('sendHits(PageView) - failed because following required attribute "pageTitle" is missing...');
+                    return null;
+                }
+                return {
+                    t: 'PAGEVIEW',
                     dl: documentLocation,
                     pt: pageTitle
                 };