+++ src/class/flagshipVisitor/types.ts
@@ -27,6 +27,7 @@ export type DecisionApiResponseData = {
 
 export type GetModificationInfoOutput = {
     campaignId: string;
+    isReference: boolean;
     variationId: string;
     variationGroupId: string;
 };
@@ -97,8 +98,9 @@ export type ModificationsInternalStatus = {
 };
 
 export type HitShape =
-    | { type: 'Screen'; data: ScreenHit } // Deprecated type: 'Screen'
-    | { type: 'ScreenView'; data: ScreenHit }
+    | { type: 'Screen'; data: ScreenViewHit } // Deprecated type: 'Screen'
+    | { type: 'ScreenView'; data: ScreenViewHit }
+    | { type: 'PageView'; data: PageViewHit }
     | { type: 'Transaction'; data: TransactionHit }
     | { type: 'Item'; data: ItemHit }
     | { type: 'Event'; data: EventHit };
@@ -138,7 +140,12 @@ export type EventHit = CommonHit & {
     pageTitle?: string;
 };
 
-export type ScreenHit = CommonHit & {
+export type ScreenViewHit = CommonHit & {
+    documentLocation: string;
+    pageTitle: string;
+};
+
+export type PageViewHit = CommonHit & {
     documentLocation: string;
     pageTitle: string;
 };